First Maker features:
- Creating a database and the directories
- Create the tables of the database
- Character creation
- Map drawing and naming (no editing for now)
- Money system settings

+++ src/Game/GameLauncher.cpp
@@ -37,7 +37,8 @@ GameLauncher::GameLauncher(int argc, char **argv)
 
     google::SetVLOGLevel("*", FLAGS_verbose);
     google::LogToStderr();
-    LOG(INFO) << "Starting GameLauncher v" << VERSION_MAJOR << "." << VERSION_MINOR << "." << VERSION_BUILD;
+    LOG(INFO) << "Starting GameLauncher v" << static_cast<short>(VERSION_MAJOR) << "." <<
+                 static_cast<short>(VERSION_MINOR) << "." << static_cast<short>(VERSION_BUILD);
     VLOG(verbosityLevel::OBJECT_CREATION) << "Creating " << className() << " => " << this;
     m_context = std::make_shared<config::Context>(argc, argv);
